requirements.txt
deepgram-sdk==3.2.3
fastapi
pymongo
pydantic
numpy
uvicorn
faiss-cpu
python-dotenv